LNT: Implement 'nt' built-in test, this is a replacement for NewNightlyTest.pl.
The usage model differs slightly from NewNightlyTest.pl. While NewNightlyTest.pl
tried to "do everything", this test just focuses on running the LLVM test-suite
in the "nightly" configuration, and collecting and reporting the data. In
particular, it expects that users have already checked out and built the
necessary projects. This makes it more adaptable, for example, to use in
buildbot, or for testing branches, usw.

It is already far more featureful than NewNightlyTest.pl. Some key ones:
 - Support a few well chosen options to allow very fast (<1s) testing of the
   script itself. This is useful for developing the LNT infrastructure. See 'lnt
   runtest nt --help'.

 - Is documented.

 - Doesn't support arbitrary arguments to the test-suite makefiles. Instead, it
   expects everything to be given as explicit arguments so that it can be noted
   and reported properly.

 - Infers and reports various pieces of information about the compiler under test.

 - Supports reporting revision numbers.

 - Has saner defaults -- doesn't run CBE or JIT by default, doesn't blow away
   test results by default. It still needs some automatic test-directory
   rotation features.

This effectively makes the LLVM test-suite Makefiles an "implementation detail",
at least for the purposes of nightly test. This in conjunction with the fast
test turn-around and other LNT testing features should allow me to make much faster
progress on the LNT infrastructure, once everyone is migrated.

git-svn-id: https://llvm.org/svn/llvm-project/zorg/trunk@101742 91177308-0d34-0410-b5e6-96231b3b80d8