Add trunk_revision and force_checkout parameters to ClangBuilder.getClangBuildFactory()
trunk_revision is used to specify a specific revision of LLVM/Clang
trunk to check out.  It is subject to WithProperties substitution.

force_checkout is used to control whether to supply the --force
argument to "svn checkout" when checking out LLVM/Clang.  It currently
only works if trunk_revision is set.

git-svn-id: https://llvm.org/svn/llvm-project/zorg/trunk@142923 91177308-0d34-0410-b5e6-96231b3b80d8