Build gdb to test it.
Turns out just adding the relevant GDB to the PATH is insufficient. I suspect
it's some other contents (LD_LIBRARY_PATH? python libraries? not sure what path
they use) that needs to be added, and I'd be glad to but not sure it's worth
the time to figure out.

For now this does incremental rebuilds anyway (which should never change) so
the build step only takes about a minute on my private builder.


git-svn-id: https://llvm.org/svn/llvm-project/zorg/trunk@169353 91177308-0d34-0410-b5e6-96231b3b80d8