Re-enable Valgrind execution that I accidentally regressed in r163020.
Pointed out by Takumi in review.

Honestly I'm not sure what's going on here & my change (163020 combined with
this fix) will still differ from the previous valgrind buildbot behavior. It
looks like, previously, valgrind was only used for clang tests. Now it will be
used for all the tests when it is enabled, I think. We could simplify this by
not creating a separate variable for the 'clangTestArgs' and the 'llvmTestArgs'
if this is the right behavior anyway.

The phased builder might change this - since we'll be building LLVM once,
separately, then we'll just be building/testing clang separately again.

Perhaps the right solution is to revert both of these changes & simply make
"check-clang" also run the clang-tools-extra tests if they are present.


git-svn-id: https://llvm.org/svn/llvm-project/zorg/trunk@171168 91177308-0d34-0410-b5e6-96231b3b80d8