jenkins: Make LLVM_PARALLEL_LINK_JOBS configurable
We hardcoded LLVM_PARALLEL_LINK_JOBS to 1 for LTO builds since we'll
generally run out of memory trying to link things concurrently
otherwise, but in some cases we may want to be more specific. Add a
flag for that.

This also fixes another cmake config test to actually run, since it
wasn't failing after my change even though I needed to change it.

git-svn-id: https://llvm.org/svn/llvm-project/zorg/trunk@239191 91177308-0d34-0410-b5e6-96231b3b80d8