Remove lldb-x86_64-ubuntu-14.10
There is lldb-x86_64-ubuntu-14.04-cmake that the exact same thing,
so there is no test coverage lost by doing this.
The Hexagon builders are hosted on the same machine and removing
this builder will allow us to enable these builders without having
to fix the configuration of the lldb builder.



git-svn-id: https://llvm.org/svn/llvm-project/zorg/trunk@261166 91177308-0d34-0410-b5e6-96231b3b80d8