Attempt to work around libc++ test-suite issues on gribozavr4
gribozavr4 is a great bot that provides 18 threads and only runs one build
at a time, so the libc++ threading tests shouldn't be timing out, right?

Wrong (I think). I believe the frequent timeouts are caused when all 18 threads
enter certain test directories where all of the tests spawn multiple threads
and all of the tests are naturally flaky.

For this reason I've added the '--shuffle' flag to LIT. Hopefully that will
avoid the failures.


git-svn-id: https://llvm.org/svn/llvm-project/zorg/trunk@280582 91177308-0d34-0410-b5e6-96231b3b80d8