[zorg] ClangBuilder with LLD
Add the option to checkout LLD, which will run its tests during "check-all"
and allow bootstrap with LLD as well as using it by default on tests. It'll
also allow bots to use '-fuse-ld=lld' on their CFLAGS and -DLLVM_ENABLE_LLD=True
in the CMake config, to use LLD for test-suite, stage 2, etc.

Also, forcing checkout_lld=False on all current bots for safety, since the
default behaviour was requested to be True for all checkouts. I'll add a new
AArch64 LLD buildbot soon that will use this setting.



git-svn-id: https://llvm.org/svn/llvm-project/zorg/trunk@287498 91177308-0d34-0410-b5e6-96231b3b80d8