Add "#llvm-build" IRC channel: An alternative place for bots to squawk.
The #llvm IRC channel is flooded with buildbot warnings. So much so that
actual human conversation is drowned out. The obvious option is to ignore the
buildbots; but they provide useful notifications that developers might want
to receive. The solution is simply to receive them elsewhere.

This patch tells the bots to also report changes in `#llvm-build`, allowing
users to silence the versions in `#llvm`.


git-svn-id: https://llvm.org/svn/llvm-project/zorg/trunk@294780 91177308-0d34-0410-b5e6-96231b3b80d8