[Polly] Enable DeLICM for -polly-position=before-vectorizer configurations.
Enable DeLICM to get test converage for its implementations, even though
it is still experimental and not yet useful to enable by default.

We also set -polly-unprofitable-scalar-accs=0 to avoid early bail-out
when there are scalar accesses that DeLICM could remove. This is not
necessary for process-unprofitable builders that never bail-out early.

The detect-only builder is not affacted; both changes do not affect the
SCoP detection itself.

Suggested-by: Tobias Grosser <tobias.grosser@inf.ethz.ch>

git-svn-id: https://llvm.org/svn/llvm-project/zorg/trunk@295741 91177308-0d34-0410-b5e6-96231b3b80d8