[zorg] Fix no-exceptions libcxx configurations
The no-exceptions builders are missing the -DLIBCXXABI_ENABLE_EXCEPTIONS=OFF
flag (without this, only the libc++ libraries will be built without exceptions
support, libc++abi will still be built with exceptions support - this is not
a meaningful configuration).

We should also be able to get rid of the unwinder dependency from these builds
and tests. Waiting for confirmation from buildbot maintainers before going ahead
with that change - need to make sure there are no other hidden dependencies on
the unwinder (otherwise the builds could break).

Differential revision: https://reviews.llvm.org/D30290

Reviewers: rengolin

git-svn-id: https://llvm.org/svn/llvm-project/zorg/trunk@295963 91177308-0d34-0410-b5e6-96231b3b80d8