clang-native-arm-lnt-perf: Move to test-suite producer
Move the ARM performance-tracking bot to the test-suite producer instead
of the old-fashioned nt one. This allows us to submit full perf results,
and is generally a better maintained producer.

If the results turn out to be too noisy, we might want to also add
'--run-under', 'taskset -a -c 1'. However, that will probably require
some fiddling with the board since the cores are dynamically put into
offline mode and that sometimes confuses taskset.

git-svn-id: https://llvm.org/svn/llvm-project/zorg/trunk@311638 91177308-0d34-0410-b5e6-96231b3b80d8