Use check-all on the android buildbot
Build/test all of compiler-rt as a single step using check-all.
This reduces the number of steps on the buildbot page, and automatically picks up any new tests (like ubsan-minimal).

It would be great if the "unit" tests could also run through lit. This is out of scope of this change.


git-svn-id: https://llvm.org/svn/llvm-project/zorg/trunk@315521 91177308-0d34-0410-b5e6-96231b3b80d8