jenkins: Do pseudo checkouts in relay jobs
Relay jobs lacked blamelists/changelogs. This happened naturally because
we cannot communicate parameters/extra info to them to not loose the
feature of jenkins merging multiple jobs in a queue.

To get the changelists back the relay jobs are now modified to do a
pseudo checkout with all the relevant repositories (using depth: zero so
we shouldn't really spend time checking out files to disk) to get the
changelists back.

git-svn-id: https://llvm.org/svn/llvm-project/zorg/trunk@318242 91177308-0d34-0410-b5e6-96231b3b80d8