OpenMPBuilder: Fix property substitution for lit arguments
The result of WithProperties needs to be rendered for a particular
build. That is why the current string conversion doesn't work.
Solve the problem by deferring the call to WithProperties until
we have built the complete argument string so that the returned
value can be passed as-is with the command array.

Differential Revision: https://reviews.llvm.org/D39690

git-svn-id: https://llvm.org/svn/llvm-project/zorg/trunk@318726 91177308-0d34-0410-b5e6-96231b3b80d8