[zorg] Enable assertions on the linux lldb bot
Summary:
This adds support for the extra_cmake_args argument to the
LLDBScriptCommandsFactory and uses that to enable assertions on the
linux bot.

While I'm in there, I also remove the "update ndk" step -- it was a bad
idea because it's introducing android-specific code into generic code,
and we never ended up using it anyway.

Reviewers: gkistanova

Subscribers: krytarowski, emaste, lldb-commits

Differential Revision: https://reviews.llvm.org/D35356

git-svn-id: https://llvm.org/svn/llvm-project/zorg/trunk@322171 91177308-0d34-0410-b5e6-96231b3b80d8