[zorg] Switch to secure http for checking out test-suite.
There are a few files in the test-suite that appear to be flagged as files that contain viruses. As a result, some build slaves can't extract the test suite and continue to fail. We were advised over IRC that the way to work around that is to use https instead of http.

Of course, this is a workaround, but we definitely need it to get our BE lnt bot (http://lab.llvm.org:8011/builders/clang-ppc64be-linux-lnt) back online.

Patch by Nemanja Ivanovic.

Differential Revision: https://reviews.llvm.org/D44731



git-svn-id: https://llvm.org/svn/llvm-project/zorg/trunk@329138 91177308-0d34-0410-b5e6-96231b3b80d8