Disable sanitizers on clang-cmake-armv8-lld
The combination of ARM, LLD and compiler-rt has flushed out some problems
with CFI on ARM https://bugs.llvm.org/show_bug.cgi?id=38690 . Until these
can be fixed or worked around disable the sanitizer build so that the CFI
tests are not run. In an ideal world it would be possible just to disable
the CFI tests but unfortunately the only option available is the
sanitizers.



git-svn-id: https://llvm.org/svn/llvm-project/zorg/trunk@341059 91177308-0d34-0410-b5e6-96231b3b80d8