[Zorg] Reland "Fix Android sanitizer flags."
This builder uses an old version of the NDK with a just-built version
of the Clang. This builder is making use of an NDK standalone
toolchain. For older NDKs, these toolchins have the C++ stdlib headers
and libraries installed to the GNU stdlib paths.

This causes an issue when I try to update the default -stdlib for
Android targets in Clang as this builder will begin looking in the
libc++ paths when it does not have an STL available at that path.

Explicitly use `-stdlib=libstdc++` to keep the old tools working with
a new Clang until this builder is using an NDK new enough to not need
it (NDK r19).

This relands https://reviews.llvm.org/D53182.

git-svn-id: https://llvm.org/svn/llvm-project/zorg/trunk@344753 91177308-0d34-0410-b5e6-96231b3b80d8