Opt documentation bots out of toolchain check
Summary: Documentation bot clang-tools-sphinx-docs is failing when I commit D57264 because the GCC version it uses is too old (it's on 4.9.2). This doesn't really matter for now, but really you should update it. Work around the issue for now with the cmake opt-out flag, and let bot maintainers update their toolchain asynchronously.

Reviewers: gkistanova, alexfh, EricWF, gribozavr, serge-sans-paille

Subscribers: jkorous, dexonsmith

Differential Revision: https://reviews.llvm.org/D57564

git-svn-id: https://llvm.org/svn/llvm-project/zorg/trunk@352836 91177308-0d34-0410-b5e6-96231b3b80d8