Added support for git/github to LLVMBuildFactory.
Now LLVMBuildFactory supports both a "legacy mode" with SVN checkout and a git/github checkout.

This provides a transparent way of dialing with the source code in all the bots which use LLVMBuildFactory directly or indirectly through UnifiedTreeBuilder.

is_legacy_mode is True by default to keep the status quo, but later will be changed to False to migrate all the supported bots to github.

Patch by Andrei Lebedev

Differential Revision: https://reviews.llvm.org/D69046


git-svn-id: https://llvm.org/svn/llvm-project/zorg/trunk@375140 91177308-0d34-0410-b5e6-96231b3b80d8